Fixed events code
