+++ Tests/Simulation/SoilNitrogenPatch/Paddock_V1.apsimx
@@ -1,7 +1,7 @@
 {
   "$type": "Models.Core.Simulations, Models",
   "ExplorerWidth": 300,
-  "Version": 103,
+  "Version": 102,
   "ApsimVersion": "0.0.0.0",
   "Name": "Simulations",
   "Children": [
@@ -530,6 +530,7 @@
               "TramplingOn": false,
               "MaximumPropLitterMovedToSoil": 0.1,
               "PastureConsumedAtMaximumRateOfLitterRemoval": 1200.0,
+              "SpeciesCutProportions": null,
               "Name": "SimpleGrazing",
               "Children": [],
               "IncludeInDocumentation": true,
